pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "visier-sqllike-shell"
authors = [
    {name = "Visier Research & Development", email = "info@visier.com"},
]
description = """Sample Command line utitlity for executing SQL-like
queries against the Visier platform"""
readme = "README.md"
requires-python = ">=3.8"
keywords = ["sql", "repl", "shell", "visier"]
license = {file = "LICENSE"}
dynamic = ["version"]
dependencies = [
    "visier-connector>=0.9.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = {attr = "repl.__version__"}

[project.scripts]
vsql-shell = "main:main"